Move `keymanager.NewARN` to `internal/aws.NewARN`
